non individual case code removal version
added a version which does not require the use of individual codes on
each case for item removal. This was done by adding a new version of the
removeItem.php file called removeItems.php. An extra function was added
in dataManager to help with this. Finally a new version of the remove
item modal dialogue was created and the old one was placed in
removeModalProdCode.php in case it is needed to switch back. Remove item
on item page is still the old version.