User account updates
Added admin add/remove, account deleting, and account approval functions
to the admin menu

Added overview page selection for either beer or food priority with a a
tab to select.

Added a field to user account registration which determines if the user
preference is for beer or food information.

Added a session check on main page which prioritizes food or beer
content depending on user preference